feat: encapsulate all player state in `Player` class
This includes stuff such as whether they're game over yet, or what their
end state is (i.e. "Blackjack", or "Busted"). A player object is aware
of any changes made to its hand, because cards can only enter their hand
through the `addCard()` method, making it very easy to update the state
with every card drawn. It can't change outside of that. It's very
convenient that there is a single source of truth for these things, that
can be easily accessed through getter methods. No need to fuss with
inelegant data structures in the game object, trying to map states to
player present in the game.