test: edge case where phpunit cli doesn't resolve dependencies
For some reason, using the `[#depends()]` attribute doesn't work when
invoking phphunit from the command line? It does work when running the
test from my editor, but it was throwing an error when running
phpunit from the cli, because the dependent test wasn't being passed its
required parameter. So I added a null check.