feat: rework `Blackjack` to use a `Dealer`, with tests
Now player play against the dealer. The game ends when the dealer is out
of the game, either because they win, lose, or stand at 17 points. After
all players have played their turns to fruition, the final scores are
printed comparing each players' score to that of the dealer. The game
can end early if the dealer wins, loses, or stands after the opening
hand is dealt.

The tests for the `Blackjack` class are horrible, there is so much
duplication between the large amounts of mocking that are required.