chore: add license
This is a pretty arbitrary choice. I'm a little torn in general between
fully free licenses and copy left licenses. I dig the idea of 'just do
whatever', but I wouldn't want someone to just turn around and resell
open source software. MPL2.0 is rather wordy and probably not really
enforceable anyway, but I like the idea of 'it has to stay open source,
but if you use it as-is as a dependency for your own project, you can do
whatever with your own project.'